move FileHeader helpers into utilities header
