+++ six/projects/csm/tests/test_sicd_csm.cpp
@@ -28,6 +28,7 @@
 #include <six/Utilities.h>
 #include <six/sicd/ComplexXMLControl.h>
 #include <six/sicd/Utilities.h>
+#include "utilities.h"
 
 // CSM includes
 #include <RasterGM.h>
@@ -37,53 +38,6 @@
 
 namespace
 {
-std::string toString(nitf::FileSecurity security)
-{
-    const std::string str = security.getClassificationSystem().toString() +
-            security.getCodewords().toString() +
-            security.getControlAndHandling().toString() +
-            security.getReleasingInstructions().toString() +
-            security.getDeclassificationType().toString() +
-            security.getDeclassificationDate().toString() +
-            security.getDeclassificationExemption().toString() +
-            security.getDowngrade().toString() +
-            security.getDowngradeDateTime().toString() +
-            security.getClassificationText().toString() +
-            security.getClassificationAuthorityType().toString() +
-            security.getClassificationAuthority().toString() +
-            security.getClassificationReason().toString() +
-            security.getSecuritySourceDate().toString() +
-            security.getSecurityControlNumber().toString();
-
-    return str;
-}
-
-std::string toString(nitf::DESubheader subheader)
-{
-    const nitf::Uint32 subheaderFieldsLen(subheader.getSubheaderFieldsLength());
-
-    std::ostringstream ostr;
-    ostr << subheader.getFilePartType().toString()
-         << subheader.getTypeID().toString()
-         << subheader.getVersion().toString()
-         << subheader.getSecurityClass().toString()
-         << toString(subheader.getSecurityGroup())
-         << std::setw(4) << std::setfill('0') << subheaderFieldsLen;
-
-    // This tacks on XML_DATA_CONTENT when applicable
-    if (subheaderFieldsLen > 0)
-    {
-        nitf::TRE tre = subheader.getSubheaderFields();
-        for (nitf::TRE::Iterator iter = tre.begin(); iter != tre.end(); ++iter)
-        {
-            nitf::Pair pair = *iter;
-            ostr << tre.getField(pair.getKey()).toString();
-        }
-    }
-
-    return ostr.str();
-}
-
 class Test
 {
 public:
@@ -246,12 +200,13 @@ private:
         {
             std::cerr << "Converted image point > " << tolerance <<
                     " away from scpPixel\n";
+            testPassed = false;
         }
         return testPassed;
     }
 
 private:
-    static const char MODEL_NAME[];;
+    static const char MODEL_NAME[];
 
     const std::string mSicdPathname;
     const csm::Plugin& mPlugin;
