+++ six/projects/csm/tests/test_sidd_csm.cpp
@@ -35,55 +35,11 @@
 #include <Plugin.h>
 #include <NitfIsd.h>
 
+#include "utilities.h"
+
 
 namespace
 {
-std::string toString(nitf::FileSecurity security)
-{
-    const std::string str = security.getClassificationSystem().toString() +
-            security.getCodewords().toString() +
-            security.getControlAndHandling().toString() +
-            security.getReleasingInstructions().toString() +
-            security.getDeclassificationType().toString() +
-            security.getDeclassificationDate().toString() +
-            security.getDeclassificationExemption().toString() +
-            security.getDowngrade().toString() +
-            security.getDowngradeDateTime().toString() +
-            security.getClassificationText().toString() +
-            security.getClassificationAuthorityType().toString() +
-            security.getClassificationAuthority().toString() +
-            security.getClassificationReason().toString() +
-            security.getSecuritySourceDate().toString() +
-            security.getSecurityControlNumber().toString();
-
-    return str;
-}
-
-std::string toString(nitf::DESubheader subheader)
-{
-    const nitf::Uint32 subheaderFieldsLen(subheader.getSubheaderFieldsLength());
-
-    std::ostringstream ostr;
-    ostr << subheader.getFilePartType().toString()
-         << subheader.getTypeID().toString()
-         << subheader.getVersion().toString()
-         << subheader.getSecurityClass().toString()
-         << toString(subheader.getSecurityGroup())
-         << std::setw(4) << std::setfill('0') << subheaderFieldsLen;
-
-    // This tacks on XML_DATA_CONTENT when applicable
-    if (subheaderFieldsLen > 0)
-    {
-        nitf::TRE tre = subheader.getSubheaderFields();
-        for (nitf::TRE::Iterator iter = tre.begin(); iter != tre.end(); ++iter)
-        {
-            nitf::Pair pair = *iter;
-            ostr << tre.getField(pair.getKey()).toString();
-        }
-    }
-
-    return ostr.str();
-}
 
 class Test
 {
@@ -94,7 +50,7 @@ public:
         mSiddPathname(siddPathname),
         mPlugin(plugin)
     {
-        // Read in the SICD XML
+        // Read in the SIDD XML
         mXmlRegistry.addCreator(six::DataType::DERIVED,
                 new six::XMLControlCreatorT<six::sidd::DerivedXMLControl>());
 
@@ -126,7 +82,7 @@ public:
                 mReader.getRecord().getDataExtensions().getFirst().getData());
         des.setSubHeader(toString(segment.getSubheader()));
 
-        // The DES's data is just the SICD XML string
+        // The DES's data is just the SIDD XML string
         des.setData(six::toXMLString(mDerivedData.get(), &mXmlRegistry));
 
         nitfIsd.addFileDes(des);
@@ -262,7 +218,7 @@ private:
     }
 
 private:
-    static const char MODEL_NAME[];;
+    static const char MODEL_NAME[];
 
     const std::string mSiddPathname;
     const csm::Plugin& mPlugin;
@@ -304,7 +260,7 @@ int main(int argc, char** argv)
         if (argc != 2)
         {
             std::cerr << "Usage: " << sys::Path::basename(argv[0])
-                      << " <SICD pathname>\n\n";
+                      << " <SIDD pathname>\n\n";
             return 1;
         }
         sys::OS os;
