+++ six/projects/csm/wscript
@@ -1,6 +1,6 @@
 MAINTAINER         = 'adam.sylvester@mdaus.com'
 VERSION            = '1.1.2'
-MODULE_DEPS        = 'scene six.sicd six.sidd'
+MODULE_DEPS        = 'six.sicd six.sidd'
 PLUGIN             = 'CSM'
 REMOVEPLUGINPREFIX = True
